requirements.txt
tqdm
imageio
imageio[ffmpeg]
kornia
trimesh
Pillow
ninja
omegaconf
plyfile
opencv_python
opencv_contrib_python
tensorboardX
matplotlib
texttable
albumentations
timm==0.9.10
scikit-learn
torch-kmeans
albumentations
albucore
open3d
tensorboard
gdown
h5py
pytorch_lightning